% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resume}
\alias{resume}
\title{Resume Audit Study Dataset}
\format{
A data frame with 4,870 observations and 4 variables:
\describe{
  \item{firstname}{Character. First name used on the resume}
  \item{sex}{Factor. Sex of the applicant ("male" or "female")}
  \item{race}{Factor. Perceived race based on the name ("black" or "white")}
  \item{call}{Logical. Callback indicator (TRUE = received callback, FALSE = no callback)}
}
}
\source{
Bertrand, M., & Mullainathan, S. (2004). Are Emily and Greg More Employable
Than Lakisha and Jamal? A Field Experiment on Labor Market Discrimination.
\emph{American Economic Review}, 94(4), 991-1013.

Data obtained from Harvard Dataverse:
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ZX3VEV}
}
\usage{
resume
}
\description{
A dataset from Bertrand and Mullainathan (2004) studying racial discrimination
in the labor market. The researchers conducted a randomized experiment by
sending fictitious resumes with either Black-sounding or White-sounding names
to help-wanted ads in Boston and Chicago newspapers to measure the effect
of perceived race on callback rates for interviews.
}
\examples{
\dontrun{
data(resume)
summary(resume)
table(resume$race)
mean(resume$call[resume$race == "white"])  # Callback rate for white-sounding names
mean(resume$call[resume$race == "black"])  # Callback rate for black-sounding names
}
}
\keyword{datasets}
